removed authentication from UI level.
