+++ lib/enum/menu_actions.dart
@@ -1,0 +1,1 @@
+enum MenuActions { logout }
