+++ lib/services/auth/auth_service.dart
@@ -1,11 +1,17 @@
+import 'package:firebase_core/firebase_core.dart';
+import 'package:mydart/firebase_options.dart';
 import 'package:mydart/services/auth/auth_provider.dart';
 import 'package:mydart/services/auth/auth_service.dart';
 import 'package:mydart/services/auth/auth_user.dart';
+import 'package:mydart/services/auth/firebase_auth_provider.dart';
 
 class AuthService implements AuthProvider {
   final AuthProvider provider;
 
   AuthService({required this.provider});
+  factory AuthService.firebase() => AuthService(
+        provider: FirebaseAuthProvider(),
+      );
 
   @override
   Future<AuthUser> createUser({
@@ -35,4 +41,7 @@ class AuthService implements AuthProvider {
 
   @override
   Future<void> sendEmailVerification() => provider.sendEmailVerification();
+
+  @override
+  Future<void> initialize() async => provider.initialize();
 }
