+++ lib/services/auth/firebase_auth_provider.dart
@@ -1,3 +1,5 @@
+import 'package:firebase_core/firebase_core.dart';
+import 'package:mydart/firebase_options.dart';
 import 'package:mydart/services/auth/auth_user.dart';
 import 'package:mydart/services/auth/auth_provider.dart';
 import 'package:mydart/services/auth/auth_exception.dart';
@@ -93,5 +95,12 @@ class FirebaseAuthProvider implements AuthProvider {
     } else {
       throw UserNotLoggedInAuthException();
     }
+  }
+
+  @override
+  Future<void> initialize() async {
+    await Firebase.initializeApp(
+      options: DefaultFirebaseOptions.currentPlatform,
+    );
   }
 }
